% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_custom_isotopes.R
\name{custom_isotopes_example}
\alias{custom_isotopes_example}
\title{An example of how to add your own elements.}
\usage{
custom_isotopes_example()
}
\description{
This can be used, for instance, with isotopically labelled molecules.
}
