% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/IsoSpecR.R
\name{res2df}
\alias{res2df}
\title{Represents the results of IsoSpec in a user friendly way.}
\usage{
res2df(IsoSpecResults)
}
\arguments{
\item{IsoSpecResults}{The list of results from calling \code{IsoSpecify}.}
}
\value{
A \code{data.frame} containing the masses, logarithms of probability, and probability of the isotopologues.
}
\description{
ATTENTION: this function sorts the results with descending probability.
The sorting is a \code{nlog(n)} procedure which asymptotically dominates the calculations of IsoSpec and should be applied only for a good reason.
}
\examples{
res <- IsoSpecify( molecule = c(C=10,H=22,O=1), stopCondition = .9999 )
res2df(res)
}

