% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Rinterface}
\alias{Rinterface}
\title{Get the bloody solution.}
\usage{
Rinterface(isotopeNumbers, atomCounts, isotopeMasses, isotopeProbabilities,
  stopCondition, algo = 0L, tabSize = 1000L, hashSize = 1000L,
  step = 0.25)
}
\arguments{
\item{isotopeNumbers}{Interger, numbers of isotopes of elements composing the studied chemical compound.}

\item{atomCounts}{Interger, numbers of atoms of elements composing the studied chemical compound.}

\item{isotopeMasses}{Numeric, masses of isotopes of elements composing the studied chemical compound.}

\item{isotopeProbabilities}{Numeric, frequencies of isotopes of elements composing the studied chemical compound.}

\item{tabSize}{Integer, technical parameter.}

\item{hashSize}{Integer, initial size of the hash-table used.}

\item{step}{Numeric, the percentage of the percentile from the previous layer below defining the threshold for the next layer.}

\item{cutOff}{Numeric, either the probability of the optimal p-set, or the threshold value of probability below which isotopologues are trimmed.}
}
\description{
Get the bloody solution.
}

